backend/requirements.txt
flask==1.1.1
gunicorn==20.0.4
flask-mongoengine
flask-restful
flask-marshmallow
flask-bcrypt
flask-jwt-extended
hubspot-api-client==3.3.0
requests